Agentic AI Q4/Class Code/Class-6/pyproject.toml
[project]
name = "class-6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.2.4",
]
